[flue] Extract style-guide result types/schemas into lib/style-guide-results.ts
Move StyleGuideFindingFromModelSchema, StyleGuideResultFromModelSchema,
StyleGuideFinding, StyleGuideResult, and assignFindingIds out of the
workflow file into a shared lib. style-guide-review.ts re-exports the
public types; code-review-orchestrator.ts imports directly from lib.
